wip: move bridge handler out of review modal
