+++ components/modals/main-modal/main-modal.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import React, { useRef } from 'react'
+import React, { useCallback, useEffect, useRef } from 'react'
 import { useConnection } from 'wagmi'
 
 import ConnectionModal from '@/components/modals/connection-modal/connection-modal'
@@ -10,6 +10,17 @@ import { SuccessModal } from '@/components/modals/success-modal'
 import { WrapModal } from '@/components/modals/wrap-modal'
 import { FailedModal } from '../failed-modal'
 import { MainModalProps } from './main-modal.types'
+
+import { useTariAccountStore } from '@/store/account'
+import { setExceededDailyLimit, setUnwrapFailed, setUnwrapSuccess } from '@/store/bridge'
+import { useBridgeTransaction } from '@/hooks/use-bridge-transaction'
+import { useBridgeToEthereum } from '@/hooks/use-bridge-to-ethereum'
+import { DeployedChains } from '@tari-project/wxtm-bridge-contracts/deployments/index'
+import { useBridgeToTari } from '@/hooks/use-bridge-to-tari'
+import { setIsModalOpen, setModalStep } from '@/store/modal'
+
+const DAILY_LIMIT_ERROR = 'Daily wrap limit exceeded'
+const DAILY_LIMIT_ERROR_TYPE = 'Forbidden'
 
 export const MainModal = ({
   success,
@@ -22,8 +33,66 @@ export const MainModal = ({
   closeModalAction,
   type,
 }: MainModalProps) => {
-  const { isConnected } = useConnection()
   const modalRef = useRef<HTMLDivElement>(null)
+  const tariAccount = useTariAccountStore((s) => s.tariAccount)
+  const { address: ethAddress, chain, isConnected } = useConnection()
+
+  const { getUserBackendBridgeTxs } = useBridgeTransaction()
+  const { bridgeToEthereum } = useBridgeToEthereum()
+
+  const chainId = (chain?.id ?? 1) as DeployedChains
+  const { bridgeToTari, isSuccess, isError, error, isPending } = useBridgeToTari(ethAddress || '0x', chainId)
+
+  const handleBridgeToEthereum = useCallback(() => {
+    if (!amount || !ethAddress) {
+      return
+    }
+
+    bridgeToEthereum({
+      amount,
+      ethAddress: ethAddress,
+      amountAfterFee: feesData.amountAfterFee,
+    })
+      .then(async () => {
+        await getUserBackendBridgeTxs()
+        setExceededDailyLimit(false)
+      })
+      .catch((e) => {
+        console.error('[ TAPPLET-BRIDGE ] Bridge operation failed:', e)
+        const error = e as Error
+        const isLimitError =
+          error?.message?.includes(DAILY_LIMIT_ERROR_TYPE) || error?.message?.includes(DAILY_LIMIT_ERROR)
+        setExceededDailyLimit(isLimitError)
+        if (isLimitError) {
+          setIsModalOpen(false)
+        }
+      })
+  }, [amount, bridgeToEthereum, ethAddress, feesData.amountAfterFee, getUserBackendBridgeTxs])
+
+  const handleBridgeToTari = async () => {
+    if (!amount || !ethAddress || !tariAccount?.address) {
+      return
+    }
+    console.debug(`[ TAPPLET-BRIDGE ] Initiating transaction...`)
+    setModalStep(3)
+    const success = await bridgeToTari(amount, ethAddress, tariAccount.address)
+    if (success) {
+      setUnwrapFailed(false)
+    } else {
+      setUnwrapFailed(true)
+    }
+  }
+
+  useEffect(() => {
+    if (isSuccess) {
+      console.debug(`[ TAPPLET-BRIDGE ] Unwrap transaction success!`)
+      setModalStep(2)
+      setUnwrapSuccess(true)
+    } else if (isError) {
+      console.error(`[ TAPPLET-BRIDGE ] Unwrap transaction failed:`, error)
+      setUnwrapFailed(true)
+    }
+  }, [isPending, isSuccess, isError, error])
 
   if (step === 0 && isConnected) return null
 
@@ -49,6 +118,8 @@ export const MainModal = ({
           closeModalAction={closeModalAction}
           ethereumAddress={ethereumAddress}
           tariWalletAddress={tariWalletAddress}
+          handleBridgeToEthereum={handleBridgeToEthereum}
+          handleBridgeToTari={handleBridgeToTari}
           feesData={feesData}
         />
       )
