+++ components/modals/review-modal/review-modal.types.ts
@@ -3,5 +3,7 @@ import { BridgeFees } from '@/hooks/use-bridge-fees'
 
 export type ReviewModalProps = {
   closeModalAction: () => void
+  handleBridgeToEthereum: () => void
+  handleBridgeToTari: () => void
   feesData: BridgeFees
 } & Pick<MainModalProps, 'amount' | 'tariWalletAddress' | 'ethereumAddress'>
